Change "hasGameStarted" Variable From int to byte
Because This variable normally has a value from 0-2 and having it as an "int" type variable uses more memory compared to the "byte" Variable.

+++ src/main/java/net/fabricmc/example/LoadingTimer.java
@@ -6,7 +6,7 @@ public class LoadingTimer implements ModInitializer {
 	public static long timeToLoad;
 	public static final long startingTime = System.currentTimeMillis();
 	private static final long STARTINGTIME2 = startingTime;
-	public static int hasGameStarted = 0;
+	public static byte hasGameStarted = 0;
 	
 	@Override
 	public void onInitialize() {
